First working version with CAMv2 over UDP and IPv4
This commit introduces a first version of the CA Basic Service,
which is able to disseminate CAMs (v2) over a given interface,
for the time being over UDP and IPv4.
The code will be updated in the future to use BTP and GeoNetworking
and to enable also the reception of CAMs with the same CA Basic
Service implementation.