Add vehicleID and improve heading management
This commits adds a new option to set a valid stationID (i.e., what
we called here "vehicleID") in the CA Basic Service module, instead
of using a sample one (--vehicle-id).
This commit also tries to improve the heading management, when
reading this information from a GNSS device. If inconsistent values
are, for any reason, read, the corresponding field in the CAMs
will be set to HeadingValue_unavailable.
This enables the enconding of the messages even in presence of
invalid heading values.