Fix bug which was causing an unexpected termination of CAM dissemination
This commit fixes a bug which was making OCABS unexpectedly terminate
the CAM dissemination.
This was due to the usage of "SCHEDULE()", which was recursively
spawning too many threads.
Starting from this commit, a timerfd is used to periodically check
the CAM conditions, instead of SCHEDULE(), which solves the
aforementioned issue (no threads are spawned, other than the one
running the CAM check conditions loop and generating new CAMs
when needed).