Implement optional transmission of CAMs via UDP
This commit implements an additional option, which makes OCABS
send CAMs via UDP to a given address and port, in addition to
the standard-compliant dissemination leveraging raw sockets.
This new options is currently being tested and a new commit is
expected to be pushed soon if any issue is found.