First version of a serial UBX/NMEA parser
This commit introduces to OScar a first version of a serial UBX/NMEA parser, that can be used as a better and more feature-rich alternative to libgps for gathering the needed positioning and kinematic data.
The parser is still under development and more tests are needed.
libgps can still be used when specifying the new --use-gpsd command line option.