Fix bug with the --can-device option
This commit fixes a bug with the --can-device option, that, even if not specified, was causing the radarReaderThr() thread to be created, trying to use the "vcan0" CAN interface and causing a fatal error if such interface was not up and running.  Now, instead, if this option is not specified no CAN bus interface will be set up, as expected.