Fix issue preventing the dissemination of VAMs
This commit fixes an issue that may have prevented the dissemination of VAMs.
The issue was due to OScar not waiting for the GNSS Fix to be available before starting the dissemination.
Therefore, sometimes the dissemination was forced to start before a fix was available, causing an error and aborting the dissemination as a whole.