Add option to set the "wrong input threshold" for the serial parser
This commit adds a new command line option (-5) to set the "wrong input threshold" for the serial parser, i.e., the number of invalid/unrecognized consecutive bytes that can be accepted before declaring the serial connection as broken and terminating the serial parser. In presence of UBX DGB messages, a good values seems to be around 10000 (while the default value is set to 1000).