Implement ego VRU support in the web-based GUI
This commit implements the support to the visualization of the ego station in the web-based GUI when OScar is used as ETSI C-ITS stack for a VRU. When sending VAMs, now OScar will show a pedestrian icon in the GUI instead of a vehicle icon. It will also properly set the station ID that can be viewed by clicking on the icon.