Fix bug that was causing all VAMs to be sent at 10 Hz due to safe_distances
This commit fixes a bug that was causing VAMs to be always sent at 10 Hz due to safe_distances. The bug was due to the fact that the get_min_distance() function was considering the ego station as a remote one, thus always detecting a very close road user that was not there. Now the get_min_distance() excludes any entry in the LDM with a station ID equal to the one of the ego ITS-S.