Add new options to send vehicle kinematic data via UDP
This commit advances the version to 7.1 and adds a new option that can be used to send some relevant vehicle kinematic data to an external application via a UDP socket, every second. The new option is a long-only option: --send-vehdata-socket, and takes as argument the IP address and port in the format IP:port.