Restructure the logic of the Metric Supervisor
This commit restructures the logic of the Metric Supervisor: instead of having a MetricSupervisor object with references to all the Basic Services and SocketClient, now we have a single MetricSupervisor which pointer is provided to all Basic Services and to the SocketClient, so that they can "signal" the transmission and reception of each packet. This makes the MetricSupervisor easier to manage, makes it a class independent of the Basic Services implementations, and avoids a double pointer inside SocketClient. This commit also introduces a new function in utils.h, to retrieve a map of MAC:RSSI from nl80211, with the last RSSI value from each device that OScar can hear. Finally, it introduces some minor fixes.